Major refactor, performance enchancements. par_filter_adjacent still a ?
