+++ src/examples.rs
@@ -1,6 +1,6 @@
-use crate::{digraph::*, global::*};
+use crate::{digraph::*, core::*};
 use std::sync::{Arc, Weak};
-use crate::traverse::Traverse;
+
 // Flow Graph
 
 // A struct which records the maximum flow and current flow for an edge
@@ -47,7 +47,7 @@ pub fn parallel_maximum_flow_edmonds_karp(g: &FlowGraph, s: usize, t: usize) -> 
 	// it. The Traverse enum will collect the edge in the results. Otherwise we
 	// skip the edge.
 	let target = g.node(&t).unwrap();
-	let explorer = |e: &RefEdge<usize, Empty, Flow>| {
+	let explorer = |e: &ArcEdge<usize, Empty, Flow>| {
 		let flow = e.load();
 		if flow.cur < flow.max {
 			if *target == e.target() {
@@ -109,7 +109,7 @@ pub fn parallel_maximum_flow_edmonds_karp(g: &FlowGraph, s: usize, t: usize) -> 
 
 pub fn maximum_flow_edmonds_karp(g: &FlowGraph, s: usize, t: usize) -> usize {
 	let target = g.node(&t).unwrap();
-	let explorer = |e: &RefEdge<usize, Empty, Flow>| {
+	let explorer = |e: &ArcEdge<usize, Empty, Flow>| {
 		let flow = e.load();
 		if flow.cur < flow.max {
 			if *target == e.target() {
@@ -155,48 +155,49 @@ pub fn maximum_flow_edmonds_karp(g: &FlowGraph, s: usize, t: usize) -> usize {
 	max_flow
 }
 
-// Maximum flow of a directed graph using the Ford-Fulkerson method.
-pub fn maximum_flow_ford_fulkerson(g: &FlowGraph, s: usize, t: usize) -> usize {
-	let target = g.node(&t).unwrap();
-	let mut max_flow: usize = 0;
-	while let Some(b) = g.depth_first(&s,
-		|e| {
-			let flow = e.load();
-			if flow.cur < flow.max {
-				if *target == e.target() {
-					Traverse::Finish
-				} else {
-					Traverse::Include
-				}
-			} else { Traverse::Skip }})
-	{
-		let path = b.backtrack().unwrap();
-		for e in b.iter() {
-			println!("{}", e.upgrade().unwrap())
+pub fn digraph_colouring() {
+	type ChromaticGraph = Digraph<usize, usize>;
+
+	let mut g = ChromaticGraph::new();
+
+	g.insert(1, 0);
+	g.insert(2, 0);
+	g.insert(3, 0);
+	g.insert(4, 0);
+	g.connect(&1, &2,  Empty);
+	g.connect(&1, &3,  Empty);
+	// g.connect(&1, &4,  Empty);
+
+	g.connect(&2, &1,  Empty);
+	g.connect(&2, &3,  Empty);
+	// g.connect(&2, &4,  Empty);
+
+	g.connect(&3, &1,  Empty);
+	g.connect(&3, &2,  Empty);
+	// g.connect(&3, &4,  Empty);
+
+	g.connect(&4, &1,  Empty);
+	g.connect(&4, &2,  Empty);
+	// g.connect(&4, &3,  Empty);
+
+	for (_, node) in g.nodes.iter() {
+		let mut unavailable = vec![];
+		for edge in node.outbound().iter() {
+			let colour = edge.target().load();
+			unavailable.push(colour);
 		}
-		println!("\n\n");
-		let mut aug_flow = std::usize::MAX;
-		for weak in path.iter() {
-			if let Some(edge) = weak.upgrade() {
-				let flow = edge.load();
-				if flow.max - flow.cur < aug_flow {
-					aug_flow = flow.max - flow.cur;
+		let mut node_colour: usize = 0;
+		loop {
+			let check = unavailable.iter().find(|predicate| { node_colour == **predicate });
+			match check {
+				Some(_) => { },
+				None => {
+					node.store(node_colour);
+					break ;
 				}
 			}
+			node_colour += 1;
 		}
-		for weak in path.iter() {
-			if let Some(edge) = weak.upgrade() {
-				let mut flow = edge.load();
-				flow.cur += aug_flow;
-				if let Some(rev_edge) = flow.rev.upgrade() {
-					let mut rev_flow = rev_edge.load();
-					rev_flow.cur -= aug_flow;
-					edge.store(flow);
-					rev_edge.store(rev_flow);
-				}
-			}
-		}
-		max_flow += aug_flow;
 	}
-	max_flow
+	g.print_nodes();
 }
